requirements.txt
streamlit
pandas
plotly
anthropic
python-dotenv